query-service/requirements.txt
fastapi
uvicorn
google-cloud-storage
duckdb
pandas
pyarrow
